#882 Refactor /rest/tasks into separate controller
